+++ src/main/java/carracing/model/Match.java
@@ -1,20 +1,21 @@
-package step3.model;
+package carracing.model;
 
 
-import step3.present.RacingDisplayVO;
+import carracing.present.RacingDisplayVO;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Match {
     List<Racer> racers = new ArrayList<>();
+
     public void addRacer(Racer racer) {
         racers.add(racer);
     }
 
     public List<RacingDisplayVO> progressDisplay() {
         List<RacingDisplayVO> racingDisplayVOs = new ArrayList<>();
-        for(Racer racer : racers) {
+        for (Racer racer : racers) {
             racingDisplayVOs.add(new RacingDisplayVO(racer));
         }
         return racingDisplayVOs;
@@ -23,7 +24,7 @@ public class Match {
     public List<RacingDisplayVO> winnerDisplay() {
         List<RacingDisplayVO> racingDisplayVOs = new ArrayList<>();
         List<Racer> winners = findWinners(racers);
-        for(Racer winner : winners) {
+        for (Racer winner : winners) {
             racingDisplayVOs.add(new RacingDisplayVO(winner));
         }
         return racingDisplayVOs;
@@ -32,8 +33,8 @@ public class Match {
     private List<Racer> findWinners(List<Racer> racers) {
         List<Racer> winners = new ArrayList<>();
         int winnerPosition = maxPosition(racers);
-        for(Racer racer : racers) {
-            if(racer.isWinner(winnerPosition)) {
+        for (Racer racer : racers) {
+            if (racer.isWinner(winnerPosition)) {
                 winners.add(racer);
             }
         }
@@ -42,8 +43,8 @@ public class Match {
 
     private int maxPosition(List<Racer> racers) {
         int maxPosition = -1;
-        for(Racer racer :racers) {
-           maxPosition = Math.max(maxPosition, racer.finalPosition());
+        for (Racer racer : racers) {
+            maxPosition = Math.max(maxPosition, racer.finalPosition());
         }
         return maxPosition;
     }
