+++ src/main/java/carracing/present/InputPresent.java
@@ -1,4 +1,4 @@
-package step3.present;
+package carracing.present;
 
 import java.util.Arrays;
 import java.util.List;
