+++ src/main/java/carracing/present/RacingDisplayVO.java
@@ -1,6 +1,6 @@
-package step3.present;
+package carracing.present;
 
-import step3.model.Racer;
+import carracing.model.Racer;
 
 import java.util.List;
